# services/Portfolio-Backend-Service/api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Interactive API documentation for the Portfolio Backend Service.",
        "title": "Portfolio Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Faris M.",
            "url": "https://farismnrr.com",
            "email": "hello@farismnrr.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/interactions/works/{slug}": {
            "get": {
                "tags": [
                    "Interaction"
                ],
                "summary": "Get Work Metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Work Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.WorksMetadata"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Interaction"
                ],
                "summary": "Update Work Metadata",
                "responses": {}
            }
        },
        "/v1/interactions/works/{slug}/like": {
            "patch": {
                "tags": [
                    "Interaction"
                ],
                "summary": "Increment Work Like",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Work Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/v1/interactions/works/{slug}/view": {
            "patch": {
                "tags": [
                    "Interaction"
                ],
                "summary": "Increment Work View",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Work Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/v1/og/fetch": {
            "get": {
                "description": "Extract Open Graph metadata (title, description, image, etc.) from a given URL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OG"
                ],
                "summary": "Fetch OG Metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target URL to fetch metadata from",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadata fetched successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.FetchMetadataResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "URL parameter is required",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch or read metadata",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/og/proxy": {
            "get": {
                "description": "Proxy an external image URL to avoid CORS or fetch failures.",
                "tags": [
                    "OG"
                ],
                "summary": "Proxy Image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "External image URL to proxy",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image binary data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "URL parameter is required",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to proxy image",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.WorksMetadata": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "likes_count": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "views_count": {
                    "type": "integer"
                }
            }
        },
...
        "handler.FetchMetadataResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This is an example description of the page."
                },
                "image": {
                    "type": "string",
                    "example": "https://example.com/og-image.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Example Title"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "For additional error data"
                },
                "details": {
                    "description": "For validation errors"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and then your token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}